+++ build.washingtonpost.com/components/Markdown/Components/StyleConverter.js
@@ -251,8 +251,8 @@ const StyleConverter = () => {
 
     return Math.sqrt(
       Math.pow(a[0] - b[0], 2) +
-        Math.pow(a[1] - b[1], 2) +
-        Math.pow(a[2] - b[2], 2)
+      Math.pow(a[1] - b[1], 2) +
+      Math.pow(a[2] - b[2], 2)
     );
   }
 
@@ -314,8 +314,6 @@ const StyleConverter = () => {
             <InlineSVG
               cushion="none"
               path="/img/doc-icons/clipboard.svg"
-              height="16"
-              width="16"
             />
           </Icon>
           Copy
