+++ build.washingtonpost.com/package.json
@@ -76,5 +76,9 @@
   "packageManager": "npm@8.11.0",
   "optionalDependencies": {
     "@next/swc-linux-x64-gnu": "12.1.0"
+  },
+  "overrides": {
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
   }
 }
