+++ build.washingtonpost.com/pages/playroom.js
@@ -87,7 +87,7 @@ export default function Playroom({
         <Kit.Box
           ref={iframeRef}
           as="iframe"
-          sandbox="allow-scripts allow-same-origin"
+          // sandbox="allow-scripts allow-same-origin"
           src={`/playroom?code=${LZString.compressToEncodedURIComponent(
             firstRenderCode
           )}`}
