+++ build.washingtonpost.com/services/getPropsTable.js
@@ -28,6 +28,8 @@ export const getPropsTable = async (component = "icon") => {
       options
     );
 
+    console.log("componentsData", componentsData);
+
     const generatedData = componentsData.map((component) => {
       const generatedProps = Object.entries(component?.props).map(
         ([key, value]) => {
@@ -57,10 +59,10 @@ export const getPropsTable = async (component = "icon") => {
               value.defaultValue === null
                 ? "----"
                 : JSON.stringify(value.defaultValue, null, 2)
-                    .replace(/\\/g, "")
-                    .replace(/"/g, "")
-                    .replace(/({)|(})|(:)/g, "")
-                    .replace(/(value)/g, ""),
+                  .replace(/\\/g, "")
+                  .replace(/"/g, "")
+                  .replace(/({)|(})|(:)/g, "")
+                  .replace(/(value)/g, ""),
           };
         }
       );
