+++ build.washingtonpost.com/services/index.js
@@ -4,5 +4,5 @@ export * from "./getDocsListBySection";
 export * from "./getDocFromComponent";
 export * from "./getNavigation";
 export * from "./getPropsTable";
-export * from "./getPackageData";
 export * from "./getContributors";
+export * from "./getBundleSize";
