+++ package.json
@@ -166,7 +166,9 @@
     }
   },
   "overrides": {
-    "react-docgen-typescript": "^2.2.2"
+    "react-docgen-typescript": "^2.2.2",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
   },
   "packageManager": "npm@8.11.0"
 }
