+++ ui/switch/src/index.ts
@@ -1,1 +1,3 @@
 export * from "./Switch";
+export * from "./SwitchRoot";
+export * from "./SwitchThumb";
